% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dev.R
\name{add_analysis_module}
\alias{add_analysis_module}
\alias{add_cs_module}
\alias{add_ts_module}
\title{Add skeleton of analysis module}
\usage{
add_analysis_module(
  name,
  ignore = FALSE,
  open = TRUE,
  source_template = "mod_template.R",
  test_template = "test-mod_template.R"
)

add_cs_module(name, ignore = FALSE, open = TRUE)

add_ts_module(name, ignore = FALSE, open = TRUE)
}
\arguments{
\item{name}{A character of module name, i.g. "abc".}

\item{ignore}{A logical to indicate whether the produced file to be ignored
by git.}

\item{open}{A logical to indicate whether open the produced file.}

\item{source_template}{A character of R source template, default is
"mod_template.R".}

\item{test_template}{A character of R source template, default is
"test-mod_template.R".}
}
\value{
A logical vector indicating if target files created successfully.
}
\description{
This function is a wrapper of \code{\link[=add_shiny_module]{add_shiny_module()}} to add skeleton for a
analysis module.
}
\details{
\itemize{
\item \code{add_cs_module()} is a wrapper of this function for cross-sectional
analysis modules.
\item \code{add_ts_module()} is a wrapper of this function for time series
analysis modules.
}
}
\section{Functions}{
\itemize{
\item \code{add_cs_module}: Add skeleton of module for
cross-sectional analysis.

\item \code{add_ts_module}: Add skeleton of module for
time series analysis.
}}

\examples{
\dontrun{

## Create skeleton of general analysis module

# Create source/test file for module of "abc"
# ("mod_abc.R" and "test-mod_abc.R")
add_analysis_module("abc")

# Try integration test of module
abc_app()

# Make sure tests of module is OK
testthat::test_file("tests/testthat/test-mod_abc.R")

## Create skeleton of cross-section analysis module

# Create source/test file for module of "abc"
# ("mod_cs_abc.R" and "test-mod_cs_abc.R")
add_analysis_module("abc")

# Try integration test of module
cs_abc_app()

# Make sure tests of module is OK
testthat::test_file("tests/testthat/test-mod_cs_abc.R")

## Create skeleton of time series analysis module

# Create source/test file for module of "abc"
# ("mod_ts_abc.R" and "test-mod_ts_abc.R")
add_analysis_module("abc")

# Try integration test of module
ts_abc_app()

# Make sure tests of module is OK
testthat::test_file("tests/testthat/test-mod_ts_abc.R")
}
}
