% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testdata.R
\name{aggregate_tsbl_vars}
\alias{aggregate_tsbl_vars}
\alias{industry_median}
\alias{industry_mean}
\title{Aggregate tsbl_vars by new key field}
\usage{
aggregate_tsbl_vars(tsbl_vars, by, .fun = median, ...)

industry_median(tsbl_vars)

industry_mean(tsbl_vars)
}
\arguments{
\item{tsbl_vars}{A tsibble of time series vars.}

\item{by}{A character of key field.}

\item{.fun}{A function to aggregate numeric variables.}

\item{...}{Params to .fun.}
}
\value{
A tsibble aggregated by new key
}
\description{
Use new key to aggregate tsbl_vars.
}
\details{
\code{industry_median()}/\code{industry_mean()} is a wrapper of \code{aggregate_tsbl_vars()}.
}
\section{Functions}{
\itemize{
\item \code{industry_median}: Aggregate mean of tsbl_vars by industries.

\item \code{industry_mean}: Aggregate median of tsbl_vars by industries.
}}

