% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cs_cor_corrplot.R
\name{corrplot_mixed}
\alias{corrplot_mixed}
\title{Customized corrplot}
\usage{
corrplot_mixed(
  cor_matrix,
  cor_test = NULL,
  lower = "number",
  upper = "circle",
  lower.col = NULL,
  upper.col = NULL,
  order = c("original", "AOE", "FPC", "hclust", "alphabet"),
  hclust.method = c("complete", "ward", "ward.D", "ward.D2", "single", "average",
    "mcquitty", "median", "centroid"),
  addrect = NULL,
  plotCI = c("n", "square", "circle", "rect"),
  bg = "white",
  mar = c(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{cor_matrix}{Matrix, the correlation matrix to visualize.}

\item{cor_test}{A list of matrix return by \code{\link[corrplot:cor.mtest]{corrplot::cor.mtest()}} which
contains:
\itemize{
\item p:    Square matrix of size FxF with p-values as cells.
\item lowCI:    Square matrix of size FxF, each cell represents the lower part of
a confidence interval.
\item uppCI:    Square matrix of size FxF, each cell represents the upper part of
a confidence interval.
}}

\item{lower}{A Character, the visualization method for the lower triangular
correlation matrix.}

\item{upper}{A Character, the visualization method for the upper triangular
correlation matrix.}

\item{lower.col}{Passed as col parameter to the lower matrix.}

\item{upper.col}{Passed as col parameter to the upper matrix.}

\item{order}{Character, the ordering method of the correlation matrix.
"original" for original order (default).
* "AOE" for the angular order of the eigenvectors.
* "FPC" for the first principal component order.
* "hclust" for the hierarchical clustering order.
* "alphabet" for alphabetical order.
See order parameter in \code{\link[corrplot:corrplot]{corrplot::corrplot()}}.}

\item{hclust.method}{A Character, the agglomeration method to be used when
order is \code{\link[stats:hclust]{stats::hclust()}}. This should be one of "ward", "ward.D",
"ward.D2", "single", "complete", "average", "mcquitty", "median" or
"centroid".See hclust.method parameter in \code{\link[corrplot:corrplot]{corrplot::corrplot()}}.}

\item{addrect}{A Integer, the number of rectangles draws on the graph
according to the hierarchical cluster, only valid when order is \code{hclust}.
If NULL (default), then add no rectangles.}

\item{plotCI}{See the plotCI parameter in \code{\link[corrplot:corrplot]{corrplot::corrplot()}}.}

\item{bg}{The background color.}

\item{mar}{See \code{\link[graphics:par]{graphics::par()}}.}

\item{...}{Additional arguments for \code{\link[corrplot:corrplot]{corrplot::corrplot()}}'s wrappers.}
}
\value{
(Invisibly) returns a reordered correlation matrix.
}
\description{
Customized corrplot
}
\examples{
\dontrun{

# Prepare cor matrix
cor_matrix <- cor(mtcars)
cor_test <- corrplot::cor.mtest(cor_matrix)

# Plot a customized corplot wich cluster segments
corrplot_mixed(cor_matrix,
  cor_test = cor_test,
  order = "hclust", addrect = 3
)

# Plot a customized corplot wich cluster segments and CI.
corrplot_mixed(cor_matrix,
  cor_test = cor_test,
  order = "hclust", addrect = 3,
  plotCI = "rect"
)
}

}
