% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cs_analysis.R
\name{cs_analysis}
\alias{cs_analysis}
\alias{cs_analysis_ui}
\alias{cs_analysis_server}
\alias{cs_analysis_app}
\title{cs_analysis}
\usage{
cs_analysis_ui(id, debug = FALSE)

cs_analysis_server(id, tsbl_vars, debug = FALSE)

cs_analysis_app(use_online_data = FALSE, debug = on_debug())
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}

\item{debug}{A logic to enable debug or not, default is on_debug() which
returns DEBUG environment variable.}

\item{tsbl_vars}{A tsibble of vars for cross-sectional analysis.}

\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function doesn't return value.
}
}
\description{
A shiny module for cs_analysis.
}
\details{
The module is an UI for user to analyze variables in cross-section.
}
\section{Functions}{
\itemize{
\item \code{cs_analysis_ui}: UI function of cs_analysis.

\item \code{cs_analysis_server}: Server function of cs_analysis.

\item \code{cs_analysis_app}: Testing App of cs_analysis.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  cs_analysis_ui("cs_analysis_module")
)

# Call control server in App server
server <- function(input, output, session) {
  cs_analysis <- cs_analysis_server(
    "cs_analysis_module",
    tsbl_vars = reactive(tsbl_vars)
  )
}

# Run testing App for integration testing
cs_analysis_app()
}

}
