% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cs_cluster_factoextra.R
\name{cs_cluster_factoextra}
\alias{cs_cluster_factoextra}
\alias{cs_cluster_factoextra_ui}
\alias{cs_cluster_factoextra_server}
\alias{cs_cluster_factoextra_app}
\title{cs_cluster_factoextra}
\usage{
cs_cluster_factoextra_ui(id)

cs_cluster_factoextra_server(id, csbl_vars)

cs_cluster_factoextra_app(use_online_data = FALSE)
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}

\item{csbl_vars}{A tibble of vars of cross-section.}

\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function return a data frame of ...
}
}
\description{
A shiny module for cs_cluster_factoextra.
}
\details{
The module is an UI for user to display plots of clusters
by \code{\link{factoextra}} package.
}
\section{Functions}{
\itemize{
\item \code{cs_cluster_factoextra_ui}: UI function of cs_cluster_factoextra.

\item \code{cs_cluster_factoextra_server}: Server function of cs_cluster_factoextra.

\item \code{cs_cluster_factoextra_app}: Testing App of cs_cluster_factoextra.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  cs_cluster_factoextra_ui("cs_cluster_factoextra_module")
)

# Call control server in App server
server <- function(input, output, session) {
  cs_cluster_factoextra <- cs_cluster_factoextra_server(
    "cs_cluster_factoextra_module",
    csbl_vars = reactive(csbl_vars)
  )
}

# Run testing App for integration testing
cs_cluster_factoextra_app()
}

}
