% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cs_cor_GGally.R
\name{cs_cor_GGally}
\alias{cs_cor_GGally}
\alias{cs_cor_GGally_ui}
\alias{cs_cor_GGally_server}
\alias{cs_cor_GGally_app}
\title{cs_cor_GGally}
\usage{
cs_cor_GGally_ui(id)

cs_cor_GGally_server(id, csbl_vars)

cs_cor_GGally_app(use_online_data = FALSE)
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}

\item{csbl_vars}{A tibble of vars of cross-section.}

\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function doesn't return value.
}
}
\description{
A shiny module for cs_cor_GGally.
}
\details{
The module is an UI for user to display plots of correlation
by \code{\link[GGally:GGally-package]{GGally}} package.
}
\section{Functions}{
\itemize{
\item \code{cs_cor_GGally_ui}: UI function of cs_cor_GGally.

\item \code{cs_cor_GGally_server}: Server function of cs_cor_GGally.

\item \code{cs_cor_GGally_app}: Testing App of cs_cor_GGally.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  cs_cor_GGally_ui("cs_cor_GGally_module")
)

# Call control server in App server
server <- function(input, output, session) {
  cs_cor_GGally <- cs_cor_GGally_server(
    "cs_cor_GGally_module",
    csbl_vars = reactive(csbl_vars)
  )
}

# Run testing App for integration testing
cs_cor_GGally_app()
}

}
