% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cs_dist_PerformanceAnalytics.R
\name{cs_dist_PerformanceAnalytics}
\alias{cs_dist_PerformanceAnalytics}
\alias{cs_dist_PerformanceAnalytics_ui}
\alias{cs_dist_PerformanceAnalytics_server}
\alias{cs_dist_PerformanceAnalytics_app}
\title{cs_dist_PerformanceAnalytics}
\usage{
cs_dist_PerformanceAnalytics_ui(id)

cs_dist_PerformanceAnalytics_server(id, csbl_vars)

cs_dist_PerformanceAnalytics_app(use_online_data = FALSE)
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}

\item{csbl_vars}{A tibble of vars of cross-section.}

\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function doesn't return value.
}
}
\description{
A shiny module for cs_dist_PerformanceAnalytics.
}
\details{
The module is an UI for user to display plots of distribution by
PerformanceAnalytics package.
}
\section{Functions}{
\itemize{
\item \code{cs_dist_PerformanceAnalytics_ui}: UI function of
cs_dist_PerformanceAnalytics.

\item \code{cs_dist_PerformanceAnalytics_server}: Server function of
cs_dist_PerformanceAnalytics.

\item \code{cs_dist_PerformanceAnalytics_app}: Testing App of cs_dist_PerformanceAnalytics.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  cs_dist_PerformanceAnalytics_ui("cs_dist_PerformanceAnalytics_module")
)

# Call control server in App server
server <- function(input, output, session) {
  cs_dist_PerformanceAnalytics <- cs_dist_PerformanceAnalytics_server(
      "cs_dist_PerformanceAnalytics_module"
  )
}

# Run testing App for integration testing
cs_dist_PerformanceAnalytics_app()
}

}
