% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cs_missing_mice.R
\name{cs_missing_mice}
\alias{cs_missing_mice}
\alias{cs_missing_mice_ui}
\alias{cs_missing_mice_server}
\alias{cs_missing_mice_app}
\title{cs_missing_mice}
\usage{
cs_missing_mice_ui(id)

cs_missing_mice_server(id, csbl_vars)

cs_missing_mice_app(use_online_data = FALSE)
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}

\item{csbl_vars}{A tibble of vars of cross-section.}

\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function doesn't return value.
}
}
\description{
A shiny module for cs_missing_mice.
}
\details{
The module is an UI for user to display missing data by \code{\link[mice:mice]{mice}} package.
}
\section{Functions}{
\itemize{
\item \code{cs_missing_mice_ui}: UI function of cs_missing_mice.

\item \code{cs_missing_mice_server}: Server function of cs_missing_mice.

\item \code{cs_missing_mice_app}: Testing App of cs_missing_mice.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  cs_missing_mice_ui("cs_missing_mice_module")
)

# Call control server in App server
server <- function(input, output, session) {
  cs_missing_mice <- cs_missing_mice_server(
    "cs_missing_mice_module",
    csbl_vars = reactive(csbl_vars)
  )
}

# Run testing App for integration testing
cs_missing_mice_app()
}

}
