% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cs_PCA_FactoMineR.R
\name{cs_PCA_FactoMineR}
\alias{cs_PCA_FactoMineR}
\alias{cs_PCA_FactoMineR_ui}
\alias{cs_PCA_FactoMineR_server}
\alias{cs_PCA_FactoMineR_app}
\title{cs_PCA_FactoMineR}
\usage{
cs_PCA_FactoMineR_ui(id)

cs_PCA_FactoMineR_server(id, csbl_vars)

cs_PCA_FactoMineR_app(use_online_data = FALSE)
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}

\item{csbl_vars}{A tibble of vars of cross-section.}

\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function return a data frame of ...
}
}
\description{
A shiny module for cs_PCA_FactoMineR.
}
\details{
The module is an UI for user to display plots of PCA
by \code{\link[FactoMineR:FactoMineR-package]{FactoMineR}} and
\code{\link{factoextra}} package.
}
\section{Functions}{
\itemize{
\item \code{cs_PCA_FactoMineR_ui}: UI function of cs_PCA_FactoMineR.

\item \code{cs_PCA_FactoMineR_server}: Server function of cs_PCA_FactoMineR.

\item \code{cs_PCA_FactoMineR_app}: Testing App of cs_PCA_FactoMineR.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  cs_PCA_FactoMineR_ui("cs_PCA_FactoMineR_module")
)

# Call control server in App server
server <- function(input, output, session) {
  cs_PCA_FactoMineR <- cs_PCA_FactoMineR_server(
    "cs_PCA_FactoMineR_module",
    csbl_vars = reactive(csbl_vars)
  )
}

# Run testing App for integration testing
cs_PCA_FactoMineR_app()
}

}
