% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_dictionary.R
\name{data_dictionary}
\alias{data_dictionary}
\alias{data_dictionary_ui}
\alias{data_dictionary_server}
\alias{data_dictionary_app}
\title{data_dictionary}
\usage{
data_dictionary_ui(id)

data_dictionary_server(id)

data_dictionary_app()
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function return a data frame of ...
}
}
\description{
A shiny module for data_dictionary.
}
\details{
The module is an UI for user to ...
}
\section{Functions}{
\itemize{
\item \code{data_dictionary_ui}: UI function of data_dictionary.

\item \code{data_dictionary_server}: Server function of data_dictionary.

\item \code{data_dictionary_app}: Testing App of data_dictionary.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  data_dictionary_ui("data_dictionary_module")
)

# Call control server in App server
server <- function(input, output, session) {
  data_dictionary <- data_dictionary_server("data_dictionary_module")
}

# Run testing App for integration testing
data_dictionary_app()
}

}
