% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\name{debug_mode}
\alias{debug_mode}
\alias{enable_debug}
\alias{disable_debug}
\alias{on_debug}
\title{Turn on/off debug mode}
\usage{
enable_debug(quiet = FALSE)

disable_debug(quiet = FALSE)

on_debug()
}
\arguments{
\item{quiet}{A logic to show message or not.}
}
\description{
You can put some debug code in your app for debugging. When app is running in
debug mode, all debug code will be activated to show debug messages or
store data for debug, etc. When app is running in non-debug mode, all
debug code will do nothing.
}
\section{Functions}{
\itemize{
\item \code{enable_debug}: Enable debug mode.

\item \code{disable_debug}: Disable debug mode.

\item \code{on_debug}: Judge whether debug is enable or not.
}}

\seealso{
Other utils_dev: 
\code{\link{add_analysis_module}()},
\code{\link{add_shiny_module}()},
\code{\link{save_debug_data}()}
}
\concept{utils_dev}
