% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_factor_extra.R
\name{get_pca_formula}
\alias{get_pca_formula}
\title{Get equation formula for loading of PCA result}
\usage{
get_pca_formula(pca_res, pc_name_prefix = NULL, ncp = Inf, digits = 3)
}
\arguments{
\item{pca_res}{An object of class PCA \link{FactoMineR};
prcomp and princomp \link{stats}; pca, dudi \link{adea4}; epPCA \link{ExPosition}.}

\item{pc_name_prefix}{A character as name prefix of primary components in
results, e.g PC for PC1...PCn. Default NULL means use original name of
primary components.}

\item{ncp}{A number of primary components kept in the results. Default Inf means all
primary components.}

\item{digits}{How many significant digits are to be used for coefficient in
formula.}
}
\value{
A vector of Character of loading formula of Primary Components
}
\description{
Get equation formula for loading of PCA results in MathJax formats which can
be display in browser.
}
\examples{
\donttest{

# Display formula in uiOutput(ns("pca_result_formula"))
# Server function
output$pca_result_formula <- renderUI({
  pca_res <- prcomp(iris[, -5], scale = TRUE)

  # Get formula of primary components of PCA
  pca_formula <- get_pca_formula(pca_res)

  pca_formula_mathJax <- purrr::map(pca_formula, .f = helpText)

  withMathJax(
    pca_formula_mathJax
  )
})
}

}
