% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_factor_extra.R
\name{get_pca_loading}
\alias{get_pca_loading}
\title{Get loading/rotation of PCA results}
\usage{
get_pca_loading(pca_res, pc_name_prefix = NULL, ncp = Inf)
}
\arguments{
\item{pca_res}{An object of class PCA (\code{\link[FactoMineR:FactoMineR-package]{FactoMineR}});
prcomp and princomp((\code{\link[stats:stats-package]{stats}}); pca, dudi('adea4\verb{); epPCA([}ExPosition`]::\link[ExPosition:ExPosition-package]{ExPosition::ExPosition}).}

\item{pc_name_prefix}{A character as name prefix of primary components in
results, e.g PC for PC1...PCn. Default NULL means use original name of
primary components.}

\item{ncp}{A number of primary components kept in the results. Default Inf means all
primary components.}
}
\value{
A matrix of loading or rotation of PCA results.
}
\description{
Get loading/rotation of PCA results
}
\examples{
\donttest{

# Result from prcomp
pca_res_prcomp <- prcomp(iris[, -5], scale = TRUE)
pca_loading <- get_pca_loading(pca_res_prcomp)

# Result from princomp
pca_res_princomp <- princomp(scale(iris[, -5]))
pca_loading <- get_pca_loading(pca_res_princomp)

# Result from FactoMineR::PCA
pca_res_PCA <- FactoMineR::PCA(iris[, -5], ncp = ncol(iris[, -5]), graph = FALSE)
pca_loading <- get_pca_loading(pca_res_PCA)
}

}
