% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_plotly.R
\name{gg_scale_compare}
\alias{gg_scale_compare}
\title{Set scale of filling or color for comparison in plot of ggplot2}
\usage{
gg_scale_compare(
  type = c("fill", "color"),
  name = "",
  base_id = "Origin",
  compare_id = "Select",
  base_color = "seagreen",
  compare_color = "red",
  base_label = base_id,
  compare_label = compare_id
)
}
\arguments{
\item{type}{A character of setting type, "fill" for
\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual}, "color" for \code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}}.}

\item{name}{A character of name to display in title of legend box.}

\item{base_id}{A character of name for base series.}

\item{compare_id}{A character of name for comparing series.}

\item{base_color}{A character of color for base series.}

\item{compare_color}{A character of color for comparing series.}

\item{base_label}{A character of label for base series to display in legend.
default is base_id.}

\item{compare_label}{A character of label for comparing series to display
in legend. default is compare_id.}
}
\description{
This is a wrapper of \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}} and
\code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}} to set two colors for comparing series.
}
\examples{
\dontrun{

p <- ds_vars \%>\%
  ggplot(aes(x = ~QR)) +
  gg_scale_compare(
    type = "fill",
    base_id = "Origin", compare_id = "Select"
  ) +
  labs(y = NULL)

# Add original plot
p <- p +
  geom_histogram(
    aes(fill = "Origin"),
    color = NA
  )
# Add comparing plot
p <- p +
  geom_histogram(
    data = ds_vars_compare,
    aes(fill = "Select"), color = NA,
  )
}

}
\seealso{
Other stats_plotly: 
\code{\link{stats_plotly}}
}
\concept{stats_plotly}
