% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testdata.R
\name{load_csbl_vars}
\alias{load_csbl_vars}
\title{Load test data of vars of cross-section in tibble}
\usage{
load_csbl_vars(use_online_data = FALSE)
}
\arguments{
\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
A tibble of vars of cross-section.
}
\description{
Load test data of vars of cross-section in tibble
}
