% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_data.R
\name{load_data}
\alias{load_data}
\alias{load_data_ui}
\alias{load_data_server}
\alias{load_data_app}
\title{load_data}
\usage{
load_data_ui(id)

load_data_server(id)

load_data_app()
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function return a data frame of ...
}
}
\description{
A shiny module for load_data.
}
\details{
The module is an UI for user to ...
}
\section{Functions}{
\itemize{
\item \code{load_data_ui}: UI function of load_data.

\item \code{load_data_server}: Server function of load_data.

\item \code{load_data_app}: Testing App of load_data.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  load_data_ui("load_data_module")
)

# Call control server in App server
server <- function(input, output, session) {
  load_data <- load_data_server("load_data_module")
}

# Run testing App for integration testing
load_data_app()
}

}
