% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testdata.R
\name{load_factors_info}
\alias{load_factors_info}
\title{Load test data of factors_info in tibble}
\usage{
load_factors_info(use_online_data = FALSE)
}
\arguments{
\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
A tsibble of factors information.
}
\description{
Load test data of factors_info in tibble
}
