% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testdata.R
\name{load_market_return}
\alias{load_market_return}
\title{Load market return time-series in tsibble}
\usage{
load_market_return(
  use_online_data = FALSE,
  period_type = c("month", "quarter", "year")
)
}
\arguments{
\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}

\item{period_type}{A character of period type, like "month", "quarter",
"year".}
}
\value{
A tsibble of market return time-series.
}
\description{
Load market return time-series in tsibble
}
