% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testdata.R
\name{load_stock_return}
\alias{load_stock_return}
\title{Load stock return time-series in tsibble}
\usage{
load_stock_return(
  use_online_data = FALSE,
  period_type = c("month", "quarter", "year"),
  stock_codes = c("000651", "000333", "600066", "000550", "600031", "000157")
)
}
\arguments{
\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}

\item{period_type}{A character of period type, like "month", "quarter",
"year".}

\item{stock_codes}{A character vector of stock codes to load. NULL means
to load all stocks.}
}
\value{
A tsibble of stock return time-series.
}
\description{
Load stock return time-series in tsibble
}
