% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testdata.R
\name{load_tsbl_vars}
\alias{load_tsbl_vars}
\title{Load test data of vars of time-series in tsibble}
\usage{
load_tsbl_vars(use_online_data = FALSE)
}
\arguments{
\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
A tsibble of vars of time-series.
}
\description{
Load test data of vars of time-series in tsibble
}
