% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testdata.R
\name{periodize_index}
\alias{periodize_index}
\title{Transform index of tsbl_var into period date}
\usage{
periodize_index(tsbl_vars, period_field = "period")
}
\arguments{
\item{tsbl_vars}{A tsibble of time series vars.}

\item{period_field}{A character name of field which contain "quarter",
"month", "week" as period indicators. Default is "period".}
}
\value{
A tsibble with with index of periodic date.
}
\description{
Transform index of tsbl_var into period date
}
\examples{
\dontrun{
tsbl_vars_new <- periodize_index(tsbl_vars, period_field = "period")
}

}
