% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(app_module = c("explore_factor"), ...)
}
\arguments{
\item{app_module}{A character of app module to be launched.}

\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
