% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\name{save_debug_data}
\alias{save_debug_data}
\title{Save data for debug}
\usage{
save_debug_data(data, output_file, overwrite = FALSE)
}
\arguments{
\item{data}{A R obj to save.}

\item{output_file}{A character of name of output file}

\item{overwrite}{A logical to overwrite existed output_file
or not. Default FALSE means not to overwrite existed output file.}
}
\value{
A character of path of output file saved or NULL.
}
\description{
Save data in RDS file for debug if \code{\link[=on_debug]{on_debug()}} is TRUE.
}
\details{
If on_debug() is TRUE, save_debug_data will save data  and return the path
of saved file, otherwise it will save nothing and return NULL.

The saved data is located in app/temp directory with name like
\emph{"outputfile_(YYYY-MM-DD).RDS"}. It will overwrite existed file if
overwrite is TRUE, otherwise it will create a new file with a name with serial
numbers, like \emph{"outputfile_(YYYY-MM-DD)_1.RDS"}.
}
\examples{
\dontrun{

# Step1, put save_debug_data() at lines where to save data for debug
result <- foo()
save_debug_data(result, output_file = "result_data")

# Step2, activate debug mode if you want to store data for debug
enable_debug()

# Step3, deactivate debug mode if you want run app normally
disable_debug()

}
}
\seealso{
Other utils_dev: 
\code{\link{add_analysis_module}()},
\code{\link{add_shiny_module}()},
\code{\link{debug_mode}}
}
\concept{utils_dev}
