% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_slice_tsbl.R
\name{slice_tsbl}
\alias{slice_tsbl}
\alias{slice_tsbl_ui}
\alias{slice_tsbl_server}
\alias{slice_tsbl_app}
\title{Slice tibble of time series by user inputs}
\usage{
slice_tsbl_ui(id, debug = FALSE)

slice_tsbl_server(id, tsbl_vars, debug = FALSE)

slice_tsbl_app()
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}

\item{debug}{A logic to enable/disable output for debug. Default FALSE
means to disable output of debug.}

\item{tsbl_vars}{A tsibble of time series to be sliced.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function return a tsibble sliced by user's filter conditions.
}
}
\description{
A shiny module to slice tibble of time series(tsibble)
by user's filter.
conditions.
}
\details{
The module is an UI for user to choose some conditions for slice(filter)
tibble of time series.

User will promote to use following filter conditions to filter:
\itemize{
\item \strong{Industry}: one or many industry codes.
\item \strong{Stock Code}: one or many stock codes.
\item \strong{Focus Variable}:  one or many variables to focus.
\item \strong{Date Range}:
\itemize{
\item Single Period:  specify report date.
\item Multiple Periods: specify start/end date.
}
}
}
\section{Functions}{
\itemize{
\item \code{slice_tsbl_ui}: UI function of slicing tibble of time series.

\item \code{slice_tsbl_server}: Server function of slicing tibble of time series.

\item \code{slice_tsbl_app}: Testing App for slicing tibble of time series.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  slice_tsbl_ui("slice_tsbl_module")
)

# Call control server in App server
server <- function(input, output, session) {
  load_factors <- slice_tsbl_server("slice_tsbl_module",
    tsbl_vars = reactive(tsbl_vars)
  )
}

# Run testing App for integration testing
slice_tsbl_app(tsbl_vars)
}

}
