% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_plotly.R
\name{stats_plotly}
\alias{stats_plotly}
\alias{freqbar_plotly}
\alias{boxplot_plotly}
\alias{hist_plotly}
\alias{density_plotly}
\alias{qqplot_plotly}
\alias{scatter_plotly}
\alias{combochart_plotly}
\title{Interactively plot stats chart by plotly}
\usage{
freqbar_plotly(
  ds_vars,
  var_name,
  ds_vars_compare = NULL,
  plot_method = c("plot_ly", "ggplot"),
  source_id = paste0("freqbar_", var_name)
)

boxplot_plotly(
  ds_vars,
  var_name,
  ds_vars_compare = NULL,
  plot_method = c("plot_ly", "ggplot"),
  source_id = paste0("boxplot_", var_name)
)

hist_plotly(
  ds_vars,
  var_name,
  ds_vars_compare = NULL,
  plot_method = c("plot_ly", "ggplot"),
  source_id = paste0("hist_", var_name)
)

density_plotly(
  ds_vars,
  var_name,
  ds_vars_compare = NULL,
  plot_method = c("plot_ly", "ggplot"),
  source_id = paste0("density_", var_name)
)

qqplot_plotly(
  ds_vars,
  var_name,
  ds_vars_compare = NULL,
  plot_method = c("plot_ly", "ggplot"),
  source_id = paste0("qqplot_", var_name)
)

scatter_plotly(
  ds_vars,
  x_var_name,
  y_var_name,
  ds_vars_compare = NULL,
  plot_method = c("plot_ly", "ggplot"),
  source_id = paste0("scatter_", x_var_name, "_", y_var_name)
)

combochart_plotly(
  ds_vars,
  continuous_var_name,
  discrete_var_name,
  ds_vars_compare = NULL,
  top_levels = 5,
  geom_type = c("boxplot", "bar_mean", "bar_median"),
  plot_method = c("plot_ly", "ggplot"),
  source_id = paste0("scatter_", continuous_var_name, "_", discrete_var_name)
)
}
\arguments{
\item{ds_vars}{A data.frame of data for plotting.}

\item{var_name}{A character of variable name.}

\item{ds_vars_compare}{A data.frame of data for plotting reference.}

\item{plot_method}{A character of method to plotting, "plot_ly" means
to use plotly function, "ggplot" means to use [\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}] +
\code{\link[plotly:ggplotly]{plotly::ggplotly()}}.}

\item{source_id}{a character string of length 1. Match the value of
this string with the source argument in \code{\link[plotly:event_data]{plotly::event_data()}} to retrieve
the event data corresponding to a specific plot
(shiny apps can have multiple plots).}

\item{x_var_name}{A character of name of variable x.}

\item{y_var_name}{A character of name of variable y.}

\item{continuous_var_name}{A character of name of continuous variable.}

\item{discrete_var_name}{A character of name of discrete variable.}

\item{top_levels}{A integer of top levels to display, default 5 means levels
with top 5 most frequencies will be dispaly.}

\item{geom_type}{A character of geom type to plot continuous variable, e.g
"boxplot", "bar_mean", "bar_median", default boxplot means use boxplot.}
}
\value{
A plotly ojbect.
}
\description{
Plot statistic chart interactively by \code{\link[plotly:plotly]{plotly}} package,
an (MIT licensed) web-based interactive charting library.
}
\section{Functions}{
\itemize{
\item \code{freqbar_plotly}: frequency bar chart for a discrete variable
in data frame.

\item \code{boxplot_plotly}: boxplot for a continuous variable in data frame.

\item \code{hist_plotly}: histogram for a continuous variable in data frame.

\item \code{density_plotly}: density plot for a continuous variable in data frame.

\item \code{qqplot_plotly}: qq plot for a continuous variable in data frame.

\item \code{scatter_plotly}: scatter plot for two continuous variable in data frame.

\item \code{combochart_plotly}: combo chart for a continuous variable and a discrete
in data frame.
}}

\seealso{
Other stats_plotly: 
\code{\link{gg_scale_compare}()}
}
\concept{stats_plotly}
