% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testdata.R
\name{tidy_tsbl}
\alias{tidy_tsbl}
\title{Tidy tsibble of variables by filling gaps and imputing nas}
\usage{
tidy_tsbl(
  tsbl_vars,
  fill_gaps = c("individual", "full", "start", "end"),
  fill_nas = c("down", "up", "downup", "updown", "keep")
)
}
\arguments{
\item{tsbl_vars}{A tsibble of time series vars.}

\item{fill_gaps}{A character of handling with gaps, default is individual.
\itemize{
\item \code{individual} inserts \code{NA} for each keyed unit within its own period.
\item \code{full} fills \code{NA} over the entire time span of the data (a.k.a. fully balanced panel).
\item \code{start} pad \code{NA} to the same starting point (i.e. \verb{min(<index>)}) across units.
\item \code{end} pad \code{NA} to the same ending point (i.e. \verb{max(<index>)}) across units.
}}

\item{fill_nas}{A character of direction in which to fill missing value,
default is down. Currently either "down" (the default), "up",
"downup" (i.e. first down and then up) or
"updown" (first up and then down). "keep" means to keep nas.}
}
\value{
A tsibble with after filling gaps and nas.
}
\description{
This function is a wrapper of \code{\link[tsibble:fill_gaps]{tsibble::fill_gaps()}} and \code{\link[tidyr:fill]{tidyr::fill()}}. It
firstly fills gas with nas, then imputes nas with front/behind values
or keeps nas.
}
\examples{
\dontrun{
   tsbl_vars_new <- periodize_index(tsbl_vars, period_field = "period")
   tidy_tsbl <- tidy_tsbl(tsbl_vars_new,
                          fill_gaps = "individual",
                          fill_nas = "down")
}
}
