% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ts_feat_basic_tidyverts.R
\name{ts_feat_basic_tidyverts}
\alias{ts_feat_basic_tidyverts}
\alias{ts_feat_basic_tidyverts_ui}
\alias{ts_feat_basic_tidyverts_server}
\alias{ts_feat_basic_tidyverts_app}
\title{ts_feat_basic_tidyverts}
\usage{
ts_feat_basic_tidyverts_ui(id)

ts_feat_basic_tidyverts_server(id, tsbl_vars, tsbl_vars_average)

ts_feat_basic_tidyverts_app(use_online_data = FALSE)
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}

\item{tsbl_vars}{A tsibble of vars of time series.}

\item{tsbl_vars_average}{A tsibble of average of vars of time series.}

\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function doesn't return value.
}
}
\description{
A shiny module for ts_feat_basic_tidyverts.
}
\details{
The module is an UI for user to display simple features of time series
tidyverts family packages.
}
\section{Functions}{
\itemize{
\item \code{ts_feat_basic_tidyverts_ui}: UI function of ts_feat_basic_tidyverts.

\item \code{ts_feat_basic_tidyverts_server}: Server function of ts_feat_basic_tidyverts.

\item \code{ts_feat_basic_tidyverts_app}: Testing App of ts_feat_basic_tidyverts.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  ts_feat_basic_tidyverts_ui("ts_feat_basic_tidyverts_module")
)

# Call control server in App server
server <- function(input, output, session) {
  ts_feat_basic_tidyverts <- ts_feat_basic_tidyverts_server(
    "ts_feat_basic_tidyverts_module",
    tsbl_vars = reactive(tsbl_vars),
    tsbl_vars_average = reactive(tsbl_vars_average)
  )
}

# Run testing App for integration testing
ts_feat_basic_tidyverts_app()
}

}
