% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testdata.R
\name{tsbl2csbl}
\alias{tsbl2csbl}
\title{Convert tsbl of vars to csbl of vars}
\usage{
tsbl2csbl(tsbl_vars)
}
\arguments{
\item{tsbl_vars}{A tsibble of time series vars.}
}
\value{
A tibble of vars of cross-section.
}
\description{
Convert vars of time-series into vars of cross-section.
}
