% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstexplorer-package.R
\docType{package}
\name{zstexplorer-package}
\alias{zstexplorer}
\alias{zstexplorer-package}
\title{zstexplorer: Explorer for China Stock Market Investment}
\description{
An interactive tool to explore data of China stock market.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chriszheng2016/zstexplorer}
  \item Report bugs at \url{https://github.com/chriszheng2016/zstexplorer/issues}
}

}
\author{
\strong{Maintainer}: Chris Zheng \email{chriszheng@vip.sina.com.cn}

}
\keyword{internal}
