% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_slice_tsbl.R
\name{slice_tsbl}
\alias{slice_tsbl}
\alias{slice_tsbl_ui}
\alias{slice_tsbl_server}
\alias{slice_tsbl_app}
\title{Slice tibble of time series by user inputs}
\usage{
slice_tsbl_ui(id, debug = FALSE)

slice_tsbl_server(
  id,
  tsbl_vars,
  slice_type = c("cross_section", "time_series"),
  debug = FALSE
)

slice_tsbl_app(
  use_online_data = FALSE,
  slice_type = c("cross_section", "time_series"),
  debug = on_debug()
)
}
\arguments{
\item{id}{An ID string of module to connecting UI function and Server
function.}

\item{debug}{A logic to enable debug or not, default is on_debug() which
returns DEBUG environment variable.}

\item{tsbl_vars}{A tsibble of time series to be sliced.}

\item{slice_type}{A character indicator about how to slice data, e.g.,
"cross_section", "time_series". Default is cross_section.}

\item{use_online_data}{A logical to determine whether to use test data from
database or not. Default FALSE means to use achieved data for tests.}
}
\value{
\itemize{
\item UI function doesn't return value.
}

\itemize{
\item Server function return a list of tsibbles sliced by
user's filter conditions, which contains following fields:
\itemize{
\item slice_vars : a tsibble of variables sliced by user.
\item slice_vars_average: a tsibble of average of variables sliced by user.
}
}
}
\description{
A shiny module to slice tibble of time series(tsibble)
by user's filter.
conditions.
}
\details{
The module is an UI for user to choose some conditions for slice(filter)
tibble of time series.

User will promote to use following filter conditions to filter:
\itemize{
\item \strong{Industry codes}: one or many industry codes.
\item \strong{Stock codes}: one or many stock codes.
\item \strong{Variable coes}:  one or many variables to focus.
\item \strong{Period}: Period type of time series, such as "quarter", "month",
"week", etc.
\item \strong{Date range}:
\itemize{
\item Single Period:  specify report date.
\item Multiple Periods: specify start/end date.
}
\item \strong{Industry average}: options for computing industry average.
\itemize{
\item Average by: a variable used to compute industry average.
\item Average method: use mean or median to compute industry average.
}
}
}
\section{Functions}{
\itemize{
\item \code{slice_tsbl_ui}: UI function of slicing tibble of time series.

\item \code{slice_tsbl_server}: Server function of slicing tibble of time series.

\item \code{slice_tsbl_app}: Testing App for slicing tibble of time series.
}}

\examples{
\dontrun{
# Set up control UI in app UI
ui <- fluidPage(
  slice_tsbl_ui("slice_tsbl_module")
)

# Call control server in App server
server <- function(input, output, session) {

  # Slice tsibble of variables by timeseries and show result in debug mode
  slice_result <- slice_tsbl_server("slice_tsbl_module",
    tsbl_vars = reactive(tsbl_vars),
    slice_type = "time_series",
    debug = TRUE
    )

 # A tsibble of variables sliced by user
 slice_tsbl_vars <- slice_result$slice_vars

 # A tsibble of industry average of variables sliced by user
 slice_vars_average <-  slice_result$slice_vars_average
}

# Run testing App for integration testing
slice_tsbl_app( use_online_data = FALSE,
                slice_type = "time_series",
}

}
